Rename 'show' module to 'demos' to fix name conflict
