+++ proplot/__init__.py
@@ -19,7 +19,7 @@ with timers._benchmark('imports'):
     from .axes import *  # noqa: F401 F403
     from .figure import *  # noqa: F401 F403
     from .ui import *  # noqa: F401 F403
-    from .show import *  # noqa: F401 F403
+    from .demos import *  # noqa: F401 F403
 
 # SCM versioning
 name = 'proplot'
