+++ proplot/demos.py
